requirements.txt
streamlit>=1.28.0
pandas>=2.0
scikit-learn>=1.3
joblib>=1.3
google-genai
requests
python-dotenv